+++ HaGame2D/Physics.h
@@ -5,11 +5,11 @@ class Physics
 {
 public:
 
-	static Vector calcForce(float mass, Vector accel) {
+	static Vector calcForce(double mass, Vector accel) {
 		return accel * mass;
 	}
 
-	static Vector calcKineticEnergy(float mass, Vector velocity) {
+	static Vector calcKineticEnergy(double mass, Vector velocity) {
 		Vector vSq = Vector(velocity.x * velocity.x, velocity.y * velocity.y, velocity.z * velocity.z);
 		return vSq * mass * 0.5;
 	}
